move subcomponents
