+++ frontend/src/components/Search/SearchContext/hooks.ts
@@ -2,8 +2,8 @@ import {
   createContext, useContext, useMemo, useState,
 } from 'react';
 import { Context } from './type';
-import getEnumKeyByString from '../../utils/getEnumKeyByString';
-import { Country, SearchSelectTypes, SearchSelectKeys } from '../../types';
+import getEnumKeyByString from '../../../utils/getEnumKeyByString';
+import { Country, SearchSelectTypes, SearchSelectKeys } from '../../../types';
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
 const SearchContext = createContext<Context>({
