+++ frontend/src/components/Search/SearchInput/SearchInput.tsx
@@ -1,8 +1,8 @@
 import { ChangeEvent } from 'react';
-import useDebounce from '../../hooks/useDebounce';
+import useDebounce from '../../../hooks/useDebounce';
 import { useSearchContext } from '../SearchContext';
-import { findAll } from '../../API';
-import { INPUT_DELAY } from '../../utils/const';
+import { findAll } from '../../../API';
+import { INPUT_DELAY } from '../../../utils/const';
 
 function SearchInput() {
   const {
