+++ frontend/src/components/Search/SearchInput/index.ts
