+++ frontend/src/components/Search/SearchSuggest/index.ts
