+++ frontend/src/components/Search/SearchSuggest/type.ts
